+++ src/kube/node.ts
@@ -1,4 +1,4 @@
-import {LoggerInstance} from 'winston';
+import {Logger} from 'winston';
 import Icinga from '../icinga';
 import JSONStream from 'json-stream';
 
@@ -6,7 +6,7 @@ import JSONStream from 'json-stream';
  * kubernetes hosts
  */
 export default class Node {
-  protected logger: LoggerInstance;
+  protected logger: Logger;
   protected kubeClient;
   protected icinga: Icinga;
   protected jsonStream: JSONStream;
@@ -20,7 +20,7 @@ export default class Node {
   /**
    * kubernetes hosts
    */
-  constructor(logger: LoggerInstance, kubeClient, icinga: Icinga, jsonStream: JSONStream, options) {
+  constructor(logger: Logger, kubeClient, icinga: Icinga, jsonStream: JSONStream, options) {
     this.logger = logger;
     this.kubeClient = kubeClient;
     this.icinga = icinga;
@@ -46,7 +46,7 @@ export default class Node {
     }
 
     host = Object.assign(host, this.options.hostDefinition);
-    this.icinga.applyHost(host.host_name, host.host_name, host, this.options.hostTemplates);
+    return this.icinga.applyHost(host.host_name, host.host_name, host, this.options.hostTemplates);
   }
 
   /**
@@ -69,7 +69,7 @@ export default class Node {
           return;
         }
 
-        this.logger.debug('received kubernetes host', {object});
+        this.logger.debug('received kubernetes host resource', {object});
         if(object.object.kind !== 'Node') {
           this.logger.error('skip invalid node object', {object: object});
           return;
@@ -80,7 +80,9 @@ export default class Node {
         }
 
         if (object.type == 'ADDED') {
-          this.prepareObject(object.object);
+          this.prepareObject(object.object).catch(err => {
+            this.logger.error('failed to handle resource', {error: err});
+          });
         }
       });